langchain-models/pyproject.toml
[project]
name = "langchain-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-generativeai>=0.8.5",
    "grandalf>=0.8",
    "huggingface-hub>=0.35.3",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "numpy>=2.2.6",
    "pydantic[email]>=2.12.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "transformers>=4.57.1",
]
